made the conection with the file in lib
