+++ app/controllers/application_controller.rb
@@ -1,5 +1,6 @@
 class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
+  require_relative '../../lib/TunesTakeoutWrapper.rb'
   protect_from_forgery with: :exception
 end
