+++ app/controllers/suggestions_controller.rb
@@ -1,5 +1,10 @@
+require "#{Rails.root}/lib/charles/TunesTakeoutWrapper.rb"
+
 class SuggestionsController < ApplicationController
   def index
+    # include tunes_takeout_wrapper
+
+    @suggestions = Charles::TunesTakeoutWrapper.find(params[:user_input])
   end
 
   def favorite
