+++ app/models/user.rb
@@ -1,2 +1,2 @@
-class User < ActiveRecord::Base
+class User
 end
