--- db/migrate/20160516222513_create_users.rb
@@ -1,8 +1,0 @@
-class CreateUsers < ActiveRecord::Migration
-  def change
-    create_table :users do |t|
-
-      t.timestamps null: false
-    end
-  end
-end
