--- db/migrate/20160516222547_create_musics.rb
@@ -1,8 +1,0 @@
-class CreateMusics < ActiveRecord::Migration
-  def change
-    create_table :musics do |t|
-
-      t.timestamps null: false
-    end
-  end
-end
