--- db/migrate/20160516222558_create_foods.rb
@@ -1,8 +1,0 @@
-class CreateFoods < ActiveRecord::Migration
-  def change
-    create_table :foods do |t|
-
-      t.timestamps null: false
-    end
-  end
-end
